add color and size in order api

+++ lib/src/explore/model/trending_products_model.dart
@@ -54,7 +54,7 @@ class Products {
   Products.fromJson(Map<String, dynamic> json) {
     mainImage = json['mainImage'];
     name = json['name'];
-    price = json['price'];
+    price = json['price'].toInt();
   }
 
   Map<String, dynamic> toJson() {
